% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{binary}
\alias{binary}
\title{The Binary Dataset}
\format{
An object of class \code{data.frame} with 400 rows and 4 columns.
}
\source{
\url{https://stats.idre.ucla.edu/r/dae/logit-regression/}
}
\usage{
binary
}
\description{
A hypothetical Admission dataset.
}
\details{
The dataset is a generated dataset used to fit logistic regression models.
}
\keyword{datasets}
