% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lda_gibbs.R
\name{extract_phi}
\alias{extract_phi}
\alias{extract_theta}
\alias{extract_log_liks}
\title{Extract parameters from a LDA_Gibbs topic model}
\usage{
extract_phi(x)

extract_theta(x)

extract_log_liks(x)
}
\arguments{
\item{x}{a \code{LDA_Gibbs} topic model}
}
\description{
The functions extracts the \code{phi} and \code{theta} parameters from
a \code{LDA_Gibbs} topic model.

In addition, the functions can also be used to extract the log likelihood values
per kept iteration.
}
