% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{Hitters}
\alias{Hitters}
\title{The Hitters Dataset}
\format{
An object of class \code{data.frame} with 322 rows and 20 columns.
}
\source{
The ISLR R package.
}
\usage{
Hitters
}
\description{
Major League Baseball Data from the 1986 and 1987 seasons.
}
\details{
The dataset has been taken from the ISLR R package dataset with
the same name. Please refer to the original dataset for details.
}
\keyword{datasets}
