% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.R
\name{ll}
\alias{ll}
\title{The log-likelihood for logistic regression}
\usage{
ll(y, X, theta)
}
\arguments{
\item{y}{a vector of binary variables}

\item{X}{a design matrix with the same length as y}

\item{theta}{a vector of parameter values}
}
\description{
The function computes the log likelihood for logistic regression given a
set of parameter values \code{theta}.
}
