% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{mixture_data}
\alias{mixture_data}
\title{The Mixture Example Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 1 columns.
}
\usage{
mixture_data
}
\description{
The data consists of a matrix with 20 different values, given in
Table 8.1 in Hastie, Tibshirani and Friedman (2009)
}
\references{
Hastie, Trevor, Tibshirani, Robert, and Friedman, Jerome.
The elements of statistical learning: data mining, inference, and prediction.
2nd Edition. Springer Science & Business Media, 2009.
}
\keyword{datasets}
