% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{mnist_example}
\alias{mnist_example}
\title{MNIST example digits}
\format{
An object of class \code{list} of length 10.
}
\usage{
mnist_example
}
\description{
The first 0-9 digits of the MNIST training dataset.
}
\keyword{datasets}
