% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{pride_and_prejudice}
\alias{pride_and_prejudice}
\title{Pride and Prejudice (tokenized)}
\format{
An object of class \code{data.frame} with 122076 rows and 3 columns.

An object of class \code{data.frame} with 122076 rows and 3 columns.
}
\usage{
pride_and_prejudice

pride_and_prejudice
}
\description{
A \code{tibble} with the chapter, paragraph and word for each token.
All word has been turned to lower case and punctuation has been removed.

A \code{tibble} with the chapter, paragraph and word for each token.
All word has been turned to lower case and punctuation has been removed.
}
\seealso{
\code{prideprejudice} in the \code{janeaustenr} R package.

\code{prideprejudice} in the \code{janeaustenr} R package.
}
\keyword{datasets}
