% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{prob2_train}
\alias{prob2_train}
\alias{prob2_test}
\title{The Prob2 Dataset}
\format{
An object of class \code{data.frame} with 200 rows and 241 columns.

An object of class \code{data.frame} with 50 rows and 241 columns.
}
\source{
\url{https://stats.idre.ucla.edu/r/dae/logit-regression/}
}
\usage{
prob2_train

prob2_test
}
\description{
A simulated dataset for ridge regression.
}
\details{
The package contain a training set with 240 covariates, one dependent variable  and 200 observations.
In addition, a separately simulated test set with 50 observations is also included.
}
\keyword{datasets}
