% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycling_mdp.R
\name{recycling_mdp}
\alias{recycling_mdp}
\title{Create a MDP based on the Recycling robot example of Sutton and Barto (2017)}
\usage{
recycling_mdp(alpha, beta, r_wait, r_search)
}
\arguments{
\item{alpha}{see Sutton and Barto (2017, Example 3.7)}

\item{beta}{see Sutton and Barto (2017, Example 3.7)}

\item{r_wait}{see Sutton and Barto (2017, Example 3.7)}

\item{r_search}{see Sutton and Barto (2017, Example 3.7)}
}
\description{
Create a MDP based on the Recycling robot example of Sutton and Barto (2017)
}
