% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{The root mean squared error between y and x}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
The function computes the root mean squared error between \code{y} and \code{x}.
}
