% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Stopwords from three lexicons}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1149 rows and 2 columns.
}
\usage{
stopwords
}
\description{
A \code{data.frame} with stop words from three lexicons.
The dataset is taken from the \code{tidytext} R package.
}
\keyword{datasets}
